+++ utils/math.c
@@ -2,12 +2,12 @@
 #include "math.h"
 
 // squishes any number to between 0 and 1
-float sigmoid(float x) {
+double sigmoid(double x) {
     return (1 / (1 + exp(-x)));
 }
 
 // the derivative of the sigmoid function
-float sigmoidPrime(float x) {
+double sigmoidPrime(double x) {
     // https://towardsdatascience.com/derivative-of-the-sigmoid-function-536880cf918e
     return sigmoid(x) * (1 - sigmoid(x));
 }
